backend/requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
bcrypt==4.1.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==7.1.2
cryptography==42.0.5
ecdsa==0.19.0
fastapi==0.110.2
frozenlist==1.4.1
greenlet==3.0.3
h11==0.9.0
idna==3.7
Mako==1.3.3
MarkupSafe==2.1.5
multidict==6.0.5
mysqlclient==2.2.4
passlib==1.7.4
pyasn1==0.6.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
PyJWT==2.8.0
pyotp==2.9.0
python-decouple==3.8
python-jose==3.3.0
python-multipart==0.0.9
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.37.2
twilio==9.0.5
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.11.8
websockets==8.1
yarl==1.9.4
